+++ CHANGELOG.md
@@ -5,6 +5,24 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [Unreleased]
+
+### Added
+
+- Adding on-screen sprites once items are equipped
+- Hints between levels
+
+### Changed
+
+- Better inventory
+- setup-loop framework
+
+### Fixed
+
+- Player attack trigger
+- Item pickup issue
+
+
 ## [0.5.2] - 2025-03-23 
 
 ### Added