+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [Unreleased]
+## [0.6.0] - 2025-03-24 
 
 ### Added
 
@@ -183,4 +183,5 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 [0.5.0]: https://github.com/KDesp73/raycasting/releases/tag/v0.5.0
 [0.5.1]: https://github.com/KDesp73/raycasting/releases/tag/v0.5.1
 [0.5.2]: https://github.com/KDesp73/raycasting/releases/tag/v0.5.2
+[0.6.0]: https://github.com/KDesp73/raycasting/releases/tag/v0.6.0